Finishes Redux refactor
