+++ client/containers/App.tsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import { exec, ChildProcess } from "child_process";
 import Check from "./Check";
-import AddWord from "../components/Word/AddWord";
-import AddGlyph from "../components/Glyph/AddGlyph";
+import AddWord from "../components/Words/AddWord";
+import AddGlyph from "../components/Glyphs/AddGlyph";
 import Bot from "../components/Bot/Bot";
 import addBannedWord from "../../lib/addBannedWord";
 import addSubstitution from "../../lib/addGlyph";
